requirements.txt
PyMuPDF 
pillow
pytesseract
docx2txt
flask
python-dotenv
supabase
pinecone
langchain
langgraph
openai
anthropic
werkzeug
marshmallow
flask-cors
langchain-pinecone
langchain_experimental
langgraph
duckduckgo-search
langchain-community
langchain-tavily
faiss-cpu
groq